added notation input state
