+++ mu4/domain/notation/CMakeLists.txt
@@ -22,16 +22,19 @@ set(MODULE notation)
 find_package(Qt5 COMPONENTS Svg REQUIRED)
 
 set(MODULE_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/inotation.h
+    ${CMAKE_CURRENT_LIST_DIR}/inotationcreator.h
+    ${CMAKE_CURRENT_LIST_DIR}/inotationinputstate.h
     ${CMAKE_CURRENT_LIST_DIR}/notationdomainmodule.cpp
     ${CMAKE_CURRENT_LIST_DIR}/notationdomainmodule.h
-    ${CMAKE_CURRENT_LIST_DIR}/interfaces/inotation.h
-    ${CMAKE_CURRENT_LIST_DIR}/interfaces/inotationcreator.h
-    ${CMAKE_CURRENT_LIST_DIR}/notation.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/notation.h
-    ${CMAKE_CURRENT_LIST_DIR}/notationcreator.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/notationcreator.h
-    ${CMAKE_CURRENT_LIST_DIR}/scorecallbacks.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/scorecallbacks.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/notation.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/notation.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/notationcreator.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/notationcreator.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/scorecallbacks.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/scorecallbacks.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/notationinputstate.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/notationinputstate.h
     )
 
 set(MODULE_LINK
