+++ mu4/domain/notation/internal/notationcreator.cpp
