+++ mu4/domain/notation/internal/scorecallbacks.cpp
